requirements.txt
Django==4.2.7
django-crispy-forms==2.0
crispy-bootstrap5==0.7
python-decouple==3.8
Pillow==10.0.1
